requirements.txt
requests
python-dotenv
smtplib
signal
sys
